Extended example of alert_creation on service
Extended example of alert_creation on service with "create_alerts_and_incidents" this is a recommended option in the GUI (when editing an existing service) and offer flexibility later on down the line to merge alerts etc. Had a support ticket come up with this and that it was throwing errors further down the line.